refactor home.tsx
