+++ src/app/page.tsx
@@ -1,4 +1,4 @@
-import { Home } from '@/app/home';
+import { Home } from './pages/home';
 
 export default function Main() {
   return <Home />;
